% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj.R
\name{adj}
\alias{adj}
\alias{validate_adj}
\alias{as_adj}
\alias{is_adj}
\title{Create an adjacency list}
\usage{
adj(...)

validate_adj(x)

as_adj(x)

is_adj(x)
}
\arguments{
\item{...}{Integer vectors or a single list of integer vectors.}
}
\value{
An \code{adj} list
}
\description{
Create an adjacency list
}
\examples{
adj(1:3, NULL, 4:5)
}
