% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{adj_factor}
\alias{adj_factor}
\alias{adj_factor_int}
\title{Factor adjacency list}
\usage{
adj_factor(
  x,
  groups,
  duplicates = c("remove", "allow", "error", "warn"),
  self_loops = c("remove", "allow", "error", "warn")
)

adj_factor_int(
  x,
  groups,
  n_groups,
  duplicates = c("remove", "allow", "error", "warn"),
  self_loops = c("remove", "allow", "error", "warn")
)
}
\arguments{
\item{x}{An \code{adj} list}

\item{groups}{A vector specifying the group membership for each node in \code{x}.
\code{adj_factor()} will process this vector with \code{\link[vctrs:vec_group]{vctrs::vec_group_id()}};
\code{adj_factor_int()} expects an (1-indexed) integer vector.}

\item{duplicates}{Controls handling of duplicate neighbors. The value
\code{"warn"} warns the user; \code{"error"} throws an error; \code{"allow"} allows
duplicates, and \code{"remove"} removes duplicates silently and then sets the
corresponding attribute to \code{"error"}.}

\item{self_loops}{Controls handling of self-loops (nodes that are adjacent
to themselves). The value \code{"warn"} warns the user; \code{"error"} throws an
error; \code{"allow"} allows self-loops, and \code{"remove"} removes self-loops
silently and then sets the corresponding attribute to \code{"error"}.}

\item{n_groups}{Number of unique groups.}
}
\value{
A new \code{adj} list.
}
\description{
Computes the factor graph of a given adjacency list by a provided grouping
vector. Nodes in the same groups are merged into single nodes in the factor
graph. The resulting multi-edges and self-loops are handled according to the
specified parameters.
}
\examples{
a <- adj(konigsberg$bridge_to, ids = konigsberg$area, duplicates = "allow")
# merge two islands (A and D)
adj_factor(a, c("AD", "B", "C", "AD"))
adj_factor_int(a, c(1L, 2L, 3L, 1L), n_group = 3L, self_loops = "allow")
}
