% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{adj_zero_index}
\alias{adj_zero_index}
\title{Convert adjacency list to use zero-based indices}
\usage{
adj_zero_index(x)
}
\arguments{
\item{x}{An adjacency list.}
}
\value{
A list of integer vectors with zero-based indices.
}
\description{
Subtracts 1 from each index in the adjacency list and returns a bare list of
integer vectors, suitable for providing to C/C++ code that uses zero-based
indexing.
}
\examples{
a <- adj(konigsberg$bridge_to, ids = konigsberg$area, duplicates = "allow")
adj_zero_index(a)
}
