% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format.adj}
\alias{format.adj}
\alias{print.adj}
\title{Format and print methods for adjacency lists}
\usage{
\method{format}{adj}(x, n = 3, ...)

\method{print}{adj}(x, n = 3, ...)
}
\arguments{
\item{x}{An \code{adj} list.}

\item{n}{Maximum number of neighbors to show before truncating with an ellipsis.}

\item{...}{Ignored.}
}
\value{
A character vector representing each entry in the adjacency list.
}
\description{
Adjacency lists are printed as sets of indices for each node.
}
\examples{
a = adj(konigsberg$bridge_to, ids = konigsberg$area, duplicates = "allow")
format(a)
print(a, n = 5)
}
