% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{konigsberg}
\alias{konigsberg}
\title{The Seven Bridges of Königsberg}
\format{
\subsection{\code{konigsberg}}{

A data frame with 4 rows and 4 columns:
\describe{
\item{area}{The four land areas, A-D, as described by Euler. Area 'A'
corresponds to the central island of Kneiphof.}
\item{bridge_to}{A list column, where each entry is a character vector
listing the areas directly connected by bridges to the area in that row.}
\item{x}{The longitude of the area center, for plotting.}
\item{y}{The latitude of the area center, for plotting.}
}
}
}
\usage{
konigsberg
}
\description{
A dataset encoding the adjacency structure of seven bridges in Königsberg,
Prussia, as described by Leonhard Euler in 1736.
}
\references{
Euler, Leonhard (1741). "Solutio problematis ad geometriam situs
pertinentis". \emph{Commentarii Academiae Scientiarum Petropolitanae}: 128–140.
}
\keyword{datasets}
