% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{t.adj}
\alias{t.adj}
\title{Transpose an adjacency list}
\usage{
\method{t}{adj}(x)
}
\arguments{
\item{x}{An \code{adj} list}
}
\value{
An \code{adj} list with edges reversed.
}
\description{
Reverse the direction of edges in an adjacency list. For undirected graphs,
this is a no-op.
}
\examples{
a <- adj(2, 3, 1)
all(t(a) == adj(3, 1, 2))
}
