% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{vec_cast.adj}
\alias{vec_cast.adj}
\title{adj Casting}
\usage{
\method{vec_cast}{adj}(x, to, ...)
}
\arguments{
\item{x}{Vectors to cast.}

\item{to}{Type to cast to. If \code{NULL}, \code{x} will be returned as is.}

\item{...}{For \code{vec_cast_common()}, vectors to cast. For
\code{vec_cast()}, \code{vec_cast_default()}, and \code{vec_restore()}, these
dots are only for future extensions and should be empty.}
}
\value{
a vector of the same length, as class \code{adj} if convertible, otherwise \code{list}
}
\description{
Dispatch methods for \code{\link[vctrs:vec_cast]{vctrs::vec_cast()}}
}
\keyword{internal}
