% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce.R
\name{vec_ptype2.adj}
\alias{vec_ptype2.adj}
\title{adj Coercion}
\usage{
\method{vec_ptype2}{adj}(x, y, ...)
}
\arguments{
\item{x, y}{Vector types.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a list of the same length, as class \code{adj} if convertible, otherwise \code{list}
}
\description{
Coercion methods for \code{\link[vctrs:vec_ptype2]{vctrs::vec_ptype2()}}
}
\keyword{internal}
