% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj-package.R
\docType{package}
\name{adj-package}
\alias{adj-package}
\title{adj: Lightweight Adjacency Lists}
\description{
Provides an S3 class to represent graph adjacency lists using 'vctrs'. Allows for creation, subsetting, combining, and pretty printing of these lists. Adjacency lists can be easily converted to zero-indexed lists, which allows for easy passing of objects to low-level languages for processing.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://christophertkenny.com/adj/}
}

}
\author{
\strong{Maintainer}: Christopher T. Kenny \email{ctkenny@proton.me} (\href{https://orcid.org/0000-0002-9386-6860}{ORCID})

Authors:
\itemize{
  \item Cory McCartan \email{mccartan@psu.edu} (\href{https://orcid.org/0000-0002-6251-669X}{ORCID})
}

}
\keyword{internal}
